LPS-84119 Buy space by combining if statements

+++ portal-kernel/src/com/liferay/portal/kernel/nio/intraband/welder/fifo/FIFOUtil.java
@@ -68,10 +68,8 @@ public class FIFOUtil {
 				createFIFO(tempFIFOFile);
 			}
 			finally {
-				if (!tempFIFOFile.delete()) {
-					if (tempFIFOFile.exists()) {
-						tempFIFOFile.deleteOnExit();
-					}
+				if (!tempFIFOFile.delete() && tempFIFOFile.exists()) {
+					tempFIFOFile.deleteOnExit();
 				}
 			}
